tree-widget: ensure that the "show all" and "hide all" buttons affect all iModel content
