+++ packages/itwin/tree-widget/src/components/trees/category-tree/CategoriesTree.tsx
@@ -143,32 +143,3 @@ async function setViewType(activeView?: Viewport) {
   const viewType = view.is3d() ? "3d" : "2d";
   await Presentation.presentation.vars(RULESET_CATEGORIES.id).setString("ViewType", viewType);
 }
-
-/**
- * Toggles visibility of categories to show or hide.
- * @alpha
- */
-export async function toggleAllCategories(viewManager: ViewManager, imodel: IModelConnection, display: boolean, viewport?: Viewport, forAllViewports?: boolean, filteredProvider?: IPresentationTreeDataProvider) {
-  // istanbul ignore next
-  const activeView = viewport ?? viewManager.getFirstOpenView();
-  const ids = await getCategories(imodel, activeView, filteredProvider);
-
-  // istanbul ignore else
-  if (ids.length > 0) {
-    CategoryVisibilityHandler.enableCategory(viewManager, imodel, ids, display, forAllViewports ?? false);
-  }
-}
-
-/**
- * Gets ids of all categories or categories from filtered data provider.
- * @alpha
- */
-export async function getCategories(imodel: IModelConnection, viewport?: Viewport, filteredProvider?: IPresentationTreeDataProvider) {
-  if (filteredProvider) {
-    const nodes = await filteredProvider.getNodes();
-    return nodes.map((node) => CategoryVisibilityHandler.getInstanceIdFromTreeNodeKey(filteredProvider.getNodeKey(node)));
-  }
-
-  const categories = await loadCategoriesFromViewport(imodel, viewport);
-  return categories.map((category) => category.key);
-}
