+++ packages/itwin/tree-widget/src/components/trees/category-tree/CategoriesTreeComponent.tsx
@@ -6,8 +6,9 @@ import "./CategoriesTree.scss";
 import React, { useCallback } from "react";
 import { useActiveIModelConnection, useActiveViewport } from "@itwin/appui-react";
 import { IModelApp } from "@itwin/core-frontend";
-import { CategoryTree, getCategories, toggleAllCategories } from "./CategoriesTree";
+import { CategoryTree } from "./CategoriesTree";
 import { CategoryVisibilityHandler } from "./CategoryVisibilityHandler";
+import { getCategories, toggleAllCategories } from "../CategoriesVisibilityUtils";
 import { TreeHeaderComponent } from "../../header/TreeHeader";
 import { useTreeFilteringState } from "../../TreeFilteringState";
 import { AutoSizer } from "../../utils/AutoSizer";
